# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task company -echo- API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/images?page={Number page}&per_page={Select count elements}": {
            "get": {
                "tags": [
                    "Img"
                ],
                "summary": "Get list images",
                "description": "Get a list of images with pagination",
                "operationId": "87bd502f96fc8af6d7b30dcb7c99fafa",
                "parameters": [
                    {
                        "name": "Number page",
                        "in": "path",
                        "description": "Number page",
                        "required": true,
                        "schema": {
                            "required": [
                                "Number page"
                            ],
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Select count elements",
                        "in": "path",
                        "description": "Select count elements",
                        "required": true,
                        "schema": {
                            "required": [
                                "Select count elements"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/images": {
            "post": {
                "tags": [
                    "Img"
                ],
                "summary": "Upload file images",
                "description": "Upload file images",
                "operationId": "2c5567934ee1b1ff9dc485010d6c9e43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Errors validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/images/{id}": {
            "get": {
                "tags": [
                    "Img"
                ],
                "summary": "Get image information",
                "description": "Get image information by id",
                "operationId": "960d3cc28f6e64ff5ea873ceaba8047c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Img id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Img"
                ],
                "summary": "Delete existing image",
                "description": "Deletes an entry in the 'img', 'media', 'copy_mini_images' tables and deletes files from the server",
                "operationId": "af8176732f546444393926ab0c28cf12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Img id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/api/subscribers/{id}": {
            "get": {
                "tags": [
                    "Subscriber"
                ],
                "summary": "Get image information",
                "description": "Get subscriber information by id",
                "operationId": "5c4306643e9bb73ca92ad80ac1bfbbde",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscriber id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}